refactor: upgrade linting dependencies (#1946)
