+++ ci/nodeEngineCheck.js
@@ -35,7 +35,7 @@ class NodeEngineCheck {
 
   /**
    * Returns an array of `package.json` files under the given path and subdirectories.
-   * @param {String} [basePath] The base path for recursive directory search.
+   * @param {String} basePath The base path for recursive directory search.
    */
   async getPackageFiles(basePath = this.nodeModulesPath) {
     try {
@@ -63,29 +63,40 @@ class NodeEngineCheck {
   /**
    * Extracts and returns the node engine versions of the given package.json
    * files.
-   * @param {String[]} files The package.json files.
-   * @param {Boolean} clean Is true if packages with undefined node versions
+   * @param {Object} config
+   * @param {String[]} config.files The package.json files.
+   * @param {Boolean} config.clean Is true if packages with undefined node versions
    * should be removed from the results.
+   * @param {RegExp[]} config.ignoredFilePaths A RegExp list of ignored file paths.
    * @returns {Object[]} A list of results.
    */
-  async getNodeVersion({ files, clean = false }) {
+  async getNodeVersion({ files, ignoredFilePaths = [], clean = false }) {
 
     // Declare response
     let response = [];
 
     // For each file
     for (const file of files) {
+      const shouldFileBeIgnored = ignoredFilePaths.find((path) => path.test(file)) !== undefined;
 
-      // Get node version
-      const contentString = await fs.readFile(file, 'utf-8');
-      const contentJson = JSON.parse(contentString);
-      const version = ((contentJson || {}).engines || {}).node;
-      
-      // Add response
-      response.push({
-        file: file,
-        nodeVersion: version
-      });
+      if (shouldFileBeIgnored) {
+        continue;
+      }
+
+      try {
+        // Get node version
+        const contentString = await fs.readFile(file, 'utf-8');
+        const contentJson = JSON.parse(contentString);
+        const version = ((contentJson || {}).engines || {}).node;
+
+        // Add response
+        response.push({
+          file: file,
+          nodeVersion: version
+        });
+      } catch (err) {
+        throw `Failed to parse package at ${file} with error ${err}`;
+      }
     }
 
     // If results should be cleaned by removing undefined node versions
@@ -98,8 +109,9 @@ class NodeEngineCheck {
   /**
    * Returns the highest semver definition that satisfies all versions
    * in the given list.
-   * @param {String[]} versions The list of semver version ranges.
-   * @param {String} baseVersion The base version of which higher versions should be
+   * @param {Object} config
+   * @param {Object[]} config.versions The list of semver version ranges.
+   * @param {String} config.baseVersion The base version of which higher versions should be
    * determined; as a version (1.2.3), not a range (>=1.2.3).
    * @returns {String} The highest semver version.
    */
@@ -112,76 +124,85 @@ class NodeEngineCheck {
 
     // Sort by min version
     const sortedMinVersions = minVersions.sort((v1, v2) => semver.compare(v1.nodeMinVersion, v2.nodeMinVersion));
-    
+
     // Filter by higher versions
     const higherVersions = sortedMinVersions.filter(v => semver.gt(v.nodeMinVersion, baseVersion));
-    // console.log(`getHigherVersions: ${JSON.stringify(higherVersions)}`);
+
     return higherVersions;
   }
 
-/**
- * Returns the node version of the parent package.
- * @return {Object} The parent package info.
- */
+  /**
+   * Returns the node version of the parent package.
+   * @return {Object} The parent package info.
+   */
   async getParentVersion() {
     // Get parent package.json version
-    const version = await this.getNodeVersion({ files: [ this.packageJsonPath ], clean: true });
-    // console.log(`getParentVersion: ${JSON.stringify(version)}`);
-    return version[0];
+    const [version] = await this.getNodeVersion({ files: [ this.packageJsonPath ], clean: true });
+
+    // If parent node version could not be determined
+    if (!version) {
+      throw `Failed to determine node engine version of parent package at ${this.packageJsonPath}`;
+    }
+
+    return version;
   }
 }
 
 async function check() {
-  // Define paths
-  const nodeModulesPath = path.join(__dirname, '../node_modules');
-  const packageJsonPath = path.join(__dirname, '../package.json');
+  try {
+    // Define paths
+    const nodeModulesPath = path.join(__dirname, '../node_modules');
+    const packageJsonPath = path.join(__dirname, '../package.json');
 
-  // Create check
-  const check = new NodeEngineCheck({
-    nodeModulesPath,
-    packageJsonPath,
-  });
+    // Create check
+    const check = new NodeEngineCheck({
+      nodeModulesPath,
+      packageJsonPath,
+    });
 
-  // Get package node version of parent package
-  const parentVersion = await check.getParentVersion();
+    // Get package node version of parent package
+    const parentVersion = await check.getParentVersion();
 
-  // If parent node version could not be determined
-  if (parentVersion === undefined) {
-    core.setFailed(`Failed to determine node engine version of parent package at ${this.packageJsonPath}`);
-    return;
-  }
+    // Determine parent min version
+    const parentMinVersion = semver.minVersion(parentVersion.nodeVersion);
 
-  // Determine parent min version
-  const parentMinVersion = semver.minVersion(parentVersion.nodeVersion);
-  
-  // Get package.json files
-  const files = await check.getPackageFiles();
-  core.info(`Checking the minimum node version requirement of ${files.length} dependencies`);
+    // Get package.json files
+    const files = await check.getPackageFiles();
+    core.info(`Checking the minimum node version requirement of ${files.length} dependencies`);
 
-  // Get node versions
-  const versions = await check.getNodeVersion({ files, clean: true });
+    // List of package.json paths that should be skipped
+    const ignoredFilePaths = [
+      // Malformed package.json for testing purpose in eslint-plugin-react
+      /eslint-plugin-react\/.*malformed.*/,
+    ];
 
-  // Get are dependencies that require a higher node version than the parent package
-  const higherVersions = check.getHigherVersions({ versions, baseVersion: parentMinVersion });
+    // Get node versions
+    const versions = await check.getNodeVersion({ files, ignoredFilePaths, clean: true });
 
-  // Get highest version
-  const highestVersion = higherVersions.map(v => v.nodeMinVersion).pop();
+    // Get are dependencies that require a higher node version than the parent package
+    const higherVersions = check.getHigherVersions({ versions, baseVersion: parentMinVersion });
 
-  // If there are higher versions
-  if (higherVersions.length > 0) {
-    console.log(`\nThere are ${higherVersions.length} dependencies that require a higher node engine version than the parent package (${parentVersion.nodeVersion}):`);
-    
-    // For each dependency
-    for (const higherVersion of higherVersions) {
-      
-      // Get package name
-      const _package = higherVersion.file.split('node_modules/').pop().replace('/package.json', '');
-      console.log(`- ${_package} requires at least node ${higherVersion.nodeMinVersion} (${higherVersion.nodeVersion})`);
+    // Get highest version
+    const highestVersion = higherVersions.map(v => v.nodeMinVersion).pop();
+
+    // If there are higher versions
+    if (higherVersions.length > 0) {
+      console.log(`\nThere are ${higherVersions.length} dependencies that require a higher node engine version than the parent package (${parentVersion.nodeVersion}):`);
+
+      // For each dependency
+      for (const higherVersion of higherVersions) {
+
+        // Get package name
+        const _package = higherVersion.file.split('node_modules/').pop().replace('/package.json', '');
+        console.log(`- ${_package} requires at least node ${higherVersion.nodeMinVersion} (${higherVersion.nodeVersion})`);
+      }
+      throw `\n❌ Upgrade the node engine version in package.json to at least '${highestVersion}' to satisfy the dependencies.\n`;
     }
-    core.setFailed(`\n❌ Upgrade the node engine version in package.json to at least '${highestVersion}' to satisfy the dependencies.\n`);
-    return;
+
+    console.log(`\n✅ All dependencies satisfy the node version requirement of the parent package (${parentVersion.nodeVersion}).\n`);
+  } catch (err) {
+    core.setFailed(err);
   }
-  console.log(`\n✅ All dependencies satisfy the node version requirement of the parent package (${parentVersion.nodeVersion}).\n`);
 }
 
 check();
