+++ package.json
@@ -36,13 +36,14 @@
   ],
   "dependencies": {
     "@babel/runtime": "7.19.0",
+    "@babel/runtime-corejs3": "7.19.1",
     "bcryptjs": "2.3.0",
     "body-parser": "1.20.0",
     "commander": "9.4.0",
     "connect-flash": "0.1.1",
     "cookie-session": "2.0.0",
     "copy-to-clipboard": "3.3.2",
-    "core-js": "3.25.1",
+    "core-js": "3.25.3",
     "csurf": "1.11.0",
     "express": "4.18.1",
     "graphiql": "2.0.7",
@@ -73,28 +74,27 @@
     "typescript": "4.8.2"
   },
   "devDependencies": {
-    "@actions/core": "1.10.0",
-    "@babel/core": "7.16.0",
-    "@babel/plugin-proposal-decorators": "7.16.4",
-    "@babel/plugin-transform-regenerator": "7.16.0",
-    "@babel/plugin-transform-runtime": "7.16.4",
-    "@babel/preset-env": "7.16.4",
-    "@babel/preset-react": "7.16.0",
+    "@actions/core": "1.9.1",
+    "@babel/core": "7.19.3",
+    "@babel/eslint-parser": "7.19.1",
+    "@babel/plugin-proposal-decorators": "7.19.3",
+    "@babel/plugin-transform-runtime": "7.19.1",
+    "@babel/preset-env": "7.19.3",
+    "@babel/preset-react": "7.18.6",
     "@saithodev/semantic-release-backmerge": "2.1.2",
-    "@semantic-release/changelog": "6.0.1",
-    "@semantic-release/commit-analyzer": "9.0.2",
-    "@semantic-release/git": "10.0.1",
-    "@semantic-release/github": "8.0.6",
-    "@semantic-release/npm": "9.0.1",
-    "@semantic-release/release-notes-generator": "10.0.3",
+    "@semantic-release/changelog": "5.0.1",
+    "@semantic-release/commit-analyzer": "8.0.1",
+    "@semantic-release/git": "9.0.0",
+    "@semantic-release/github": "7.2.3",
+    "@semantic-release/npm": "7.1.3",
+    "@semantic-release/release-notes-generator": "9.0.3",
     "@types/jest": "29.1.1",
     "all-node-versions": "11.0.1",
-    "babel-eslint": "10.1.0",
-    "babel-loader": "8.2.3",
+    "babel-loader": "8.2.5",
     "css-loader": "6.7.1",
-    "eslint": "6.8.0",
-    "eslint-plugin-jest": "23.8.2",
-    "eslint-plugin-react": "7.19.0",
+    "eslint": "8.24.0",
+    "eslint-plugin-jest": "27.0.4",
+    "eslint-plugin-react": "7.31.8",
     "http-server": "14.0.0",
     "jest": "29.1.2",
     "jest-environment-jsdom": "29.1.2",
@@ -107,7 +107,7 @@
     "request-promise": "4.2.5",
     "sass": "1.55.0",
     "sass-loader": "13.0.2",
-    "semantic-release": "19.0.5",
+    "semantic-release": "17.4.6",
     "style-loader": "3.3.1",
     "svg-prep": "1.0.4",
     "webpack": "5.64.4",
