+++ src/components/Sidebar/Pin.react.js
@@ -5,7 +5,7 @@ import styles from 'components/Sidebar/Sidebar.scss';
 
 const Pin = ({ onClick }) => (
   <div className={styles.pinContainer} onClick={onClick}>
-    <Icon className={styles.sidebarPin} name="pin" width={20} height={20} />
+    <Icon className={styles.sidebarPin} name='pin' width={20} height={20} />
   </div>
 );
